Move spec tests from into smaller units
Moved from `src/test.js` to `test/*.js`. Also increased the test
suites to encaptualate fewer (mostly just one) assertions.
